feat: Split Settings into Database and Operational Subsets
